pyproject.toml
[tool.poetry]
name = "django-js-choices"
version = "0.3.1"
description = "Javascript model field choices handling for Django."
license = "MIT"
authors = ["Lorenzo Peña <lorinkoz@gmail.com>"]
readme = "README.rst"
repository = "https://github.com/lorinkoz/django-js-choices"
keywords = ["django", "choices", "javascript"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Django",
    "Framework :: Django :: 2.0",
    "Framework :: Django :: 2.1",
    "Framework :: Django :: 2.2",
    "Framework :: Django :: 3.0",
    "Framework :: Django :: 3.1",
]

[tool.poetry.dependencies]
python = "^3.6.1"
django = ">=2.0,<4.0"
rjsmin = "^1.1.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
coverage = {extras = ["toml"], version = "^5.1"}
django-multiselectfield = "^0.1.12"
dukpy = "^0.2.3"
flake8 = "^3.8.4"
flake8-bugbear = "^20.11.1"
flake8-comprehensions = "^3.3.1"
flake8-no-types = "^1.1.1"
flake8-tidy-imports = "^4.2.1"
isort = "^5.6.4"
dj-inmemorystorage = "^2.1.0"

[tool.black]
line-length = 120

[tool.isort]
atomic = true
combine_as_imports = true
line_length = 120
profile = "black"

[tool.coverage.run]
source = ["django_js_choices"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
